refactor: code health improvements for AI feature (Phase 3)
- Extract shared styled components, helpers, constants, and types from
  AISidePanel and GlobalAISidePanel into a shared module
- Consolidate duplicated backend constants into AIConstants.java
- Wrap blocking InetAddress.getByName() in Mono.fromCallable with
  boundedElastic scheduler to avoid blocking the Netty event loop
- Add non-SQL heuristics to extractReferencedTableNames for MongoDB
  and NoSQL query patterns (db.collection.method(), JSON value matching)
