+++ app/client/src/ce/components/editorComponents/GPT/shared/types.ts
@@ -1,0 +1,11 @@
+export interface QuickAction {
+  label: string;
+  icon: string;
+  prompt: string;
+}
+
+export interface CodeBlockPart {
+  type: "text" | "code";
+  content: string;
+  language?: string;
+}
