+++ app/server/appsmith-server/src/main/java/com/appsmith/server/constants/AIConstants.java
@@ -1,0 +1,12 @@
+package com.appsmith.server.constants;
+
+public final class AIConstants {
+    private AIConstants() {}
+
+    public static final String DEFAULT_AZURE_API_VERSION = "2024-12-01-preview";
+    public static final int DEFAULT_AZURE_MAX_COMPLETION_TOKENS = 16384;
+    public static final String DEFAULT_CLAUDE_MODEL = "claude-sonnet-4-6";
+    public static final String DEFAULT_CLAUDE_BASE_URL = "https://api.anthropic.com";
+    public static final String DEFAULT_OPENAI_MODEL = "gpt-4";
+    public static final String DEFAULT_OPENAI_BASE_URL = "https://api.openai.com";
+}
